services/python-ai/requirements.txt
fastapi
uvicorn
pydantic
openai-whisper
setuptools-rust
ffmpeg-python
torch
torchaudio
python-multipart
